Support multiple language game data
- Created `xiv-gen-macros` crate with `FromCsv` derive macro.
- Refactored `xiv-gen` to use manually defined structs, reducing binary size and supporting multiple languages.
- Updated `xiv-gen-db` to generate compressed bincode files for EN, JA, DE, and FR.
- Updated `ultros` backend to serve language-specific game data.
- Updated `ultros-client` to fetch game data for the user's language.
- Fixed extensive compilation errors and clippy warnings in `ultros-app` and `ultros`.
- Improved `SubrowKey` parsing and added missing fields for various game entities.
- Cleaned up unnecessary casts and other lint warnings.

+++ ultros-frontend/ultros-app/src/components/add_recipe_to_current_list.rs
@@ -25,7 +25,7 @@ pub fn AddRecipeToCurrentListModal(
     let recipe_list = StoredValue::new(
         recipes
             .values()
-            .filter_map(|r| items.get(&ItemId(r.item_result as i32)).map(|i| (i, r)))
+            .filter_map(|r| items.get(&ItemId(r.item_result)).map(|i| (i, r)))
             .collect::<Vec<_>>(),
     );
